Brings in:
- 54b9897 deps(lmdb): heed 0.11.0 -> 0.22.1, fix env-aliasing bug
  (clears ENOSPC on env.open() that blocks rholang/casper/node tests
  on macOS 26 / Darwin 25.3.0)
- 1117a04, 99a54ce: malloc_trim block-interval guards (casper, node)
- d177d35: rholang FFI i64 cast cleanup
- c3752ce: workspace-wide proptest-regressions gitignore
- ad228af: CI explicit toolchain pin
- 456c39c: CI multi-arch OCIR Docker image

Conflict resolution:
- .gitignore: kept HEAD's commented form (functionally equivalent to
  c3752ce's bare rule).
- Cargo.lock: took theirs; re-validated by 'cargo build --workspace'
  (clean).

deny.toml: ignore RUSTSEC-2026-0097 (rand unsoundness via custom logger
- not reachable in this codebase; new advisory surfaced by the heed
branch's Cargo.lock).

Verified: 'cargo test -p rholang --test abort_spec' now passes 3/3
(previously failed with ENOSPC).

Same fix already applied on staging. deps/heed-0.22-upgrade branch
will be deleted after this merge.

+++ rholang/src/lib.rs
@@ -1125,8 +1125,8 @@ extern "C" fn set_block_data(
     let params_slice = unsafe { std::slice::from_raw_parts(params_ptr, params_bytes_len) };
     let params = BlockDataProto::decode(params_slice).unwrap();
     let block_data = BlockData {
-        time_stamp: params.time_stamp as i64,
-        block_number: params.block_number as i64,
+        time_stamp: params.time_stamp,
+        block_number: params.block_number,
         sender: PublicKey::from_bytes(&params.public_key),
         seq_num: params.seq_num,
     };
